factor out tab navigation
